Separate array module from utils
